requirements.txt
Django~=5.1.8
reportlab~=4.3.1
prometheus_client~=0.21.1
djangorestframework~=3.16.0
python-dotenv~=1.1.0
asgiref~=3.8.1
social-auth-app-django~=5.4.3
drf_yasg~=1.21.10
django-tailwind~=4.0.1
django-prometheus~=2.3.1
psycopg2~=2.9.10
django_extensions~=4.1
python-docx~=1.1.2
python-telegram-bot~=22.0